Refactoring of the system. Using annotations to build objects based on tables in the bank statements. Providing a framework to make the process of addition of new bank statement readers faster.
Adding ColumnDataConverter that is used by @TableColumn annotation that must be used on parameters of constructors of DTOs. This annotation contains information of the column index and the data converter.

To work with reflection, ConstructorUtil was created.

StringUtils was renamed to BigDecimalUtils.

Added editorconfig file.

Developers must add new adapter for new bank statements and its type.